gptel-agent-tools: Set overlay priority for todo overlays
* gptel-agent-tools.el (gptel-agent--write-todo): When there is a
tool-call confirmation overlay, the todo-write overlay's keymap
takes priority, making the tool call confirmation keybindings
inaccessible.  Fix by lowering the priority of the todo-write
overlay.

Ideally we would compose the keymaps from the two overlays, but
this requires code in gptel specific to gptel-agent, which we want
to avoid.  (The code will have to be in the tool confirmation
overlay since typically that's placed second, after the todo write
overlay.)